manager can sign in on index page
