+++ backend/src/resource/page/config.html
@@ -6,12 +6,9 @@
     <link rel="stylesheet" href="/asset/purecss.min-b565cdb90b7aac9f.css">
 </head>
 <body>
-<h1>
-    博客基本信息配置
-</h1>
 <div class="pure-g">
     <div class="pure-u-1-1">
-        博客信息配置
+        <h1>博客信息配置</h1>
     </div>
 </div>
 <div class="pure-g">
