+++ backend/src/resource/page/reg.html
@@ -29,7 +29,7 @@
 <body>
 <div class="pure-g">
     <div class="pure-u-1-1">
-        创建管理员
+        <h1>创建管理员</h1>
     </div>
 </div>
 <div class="pure-g">
@@ -41,7 +41,10 @@
                 <div class="pure-control-group">
                     <label for="aligned-email">邮箱地址</label>
                     <input type="email" id="aligned-email" placeholder="邮箱地址" />
-                    <span class="pure-form-message-inline">用于管理博客或忘记密码之用.</span>
+                </div>
+                <div class="pure-control-group">
+                    <label></label>
+                    <span class="pure-form-message-inline">用于管理博客或忘记密码之用，如果忘记管理员账号，可以删除<</span>
                 </div>
                 <div class="pure-control-group">
                     <label for="aligned-password1">登录密码</label>
