+++ backend/src/service/asset_list.rs
@@ -1,6 +1,6 @@
 [
-("index-ca36d5bef1a81420.js", include_bytes!(r#"..\resource\asset/index-ca36d5bef1a81420.js.gz"#), "text/javascript"),
-("index-ca36d5bef1a81420_bg.wasm", include_bytes!(r#"..\resource\asset/index-ca36d5bef1a81420_bg.wasm.gz"#), "application/wasm"),
+("index-a6f73cf3c48f50dc.js", include_bytes!(r#"..\resource\asset/index-a6f73cf3c48f50dc.js.gz"#), "text/javascript"),
+("index-a6f73cf3c48f50dc_bg.wasm", include_bytes!(r#"..\resource\asset/index-a6f73cf3c48f50dc_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
 ("purecss.min-b565cdb90b7aac9f.css", include_bytes!(r#"..\resource\asset/purecss.min-b565cdb90b7aac9f.css.gz"#), "text/css"),
 ]
