--- frontend/asset/grid.css
@@ -1,26 +1,0 @@
-/* https://css-tricks.com/snippets/css/complete-guide-grid/ */
-.container {
-    margin: 20px;
-    display: grid;
-    grid-template-columns: 300px auto 300px;
-    grid-template-rows: auto;
-    grid-template-areas:
-    "main main main"
-    "nav footer user";
-}
-.item-nav {
-    grid-area: nav;
-}
-.item-user {
-    grid-area: user;
-    text-align: right;
-}
-.item-main {
-    grid-area: main;
-}
-.item-footer {
-    grid-area: footer;
-}
-.item-footer div {
-    text-align: center;
-}
