+++ frontend/index.html
@@ -4,6 +4,8 @@
         <meta charset="utf-8">
         <title>70年代、80年代、90年代，怀旧、复古、回忆</title>
         <link data-trunk rel="css" href="asset/purecss.min.css">
+        <link data-trunk rel="copy-file" href="resource/easy-markdown-editor-2.14.0/easymde.min.css"/>
+        <link data-trunk rel="copy-file" href="resource/easy-markdown-editor-2.14.0/easymde.min.js"/>
         <style type="text/css">
             body{padding:20px}
             .tar{text-align:right}
